+++ zarrs/benches/array_uncompressed.rs
@@ -1,10 +1,7 @@
 //! Benchmark uncompressed unsharded and sharded arrays.
 #![allow(missing_docs)]
 
-use std::sync::Arc;
-
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, Throughput};
-use zarrs::array::codec::array_to_bytes::sharding::ShardingCodecBuilder;
 
 fn array_write_all(c: &mut Criterion) {
     let mut group = c.benchmark_group("array_write_all");
@@ -39,20 +36,13 @@ fn array_write_all_sharded(c: &mut Criterion) {
         group.bench_with_input(BenchmarkId::from_parameter(size), size, |b, &size| {
             b.iter(|| {
                 let store = zarrs::storage::store::MemoryStore::new();
-                let sharding_codec = Arc::new(
-                    ShardingCodecBuilder::new(
-                        vec![32; 3].try_into().unwrap(),
-                        &zarrs::array::DataType::UInt16,
-                    )
-                    .build(),
-                );
                 let array = zarrs::array::ArrayBuilder::new(
                     vec![size; 3],
                     vec![size; 3],
                     zarrs::array::DataType::UInt16,
                     0u16,
                 )
-                .array_to_bytes_codec(sharding_codec)
+                .subchunk_shape(vec![32; 3])
                 .build(store.into(), "/")
                 .unwrap();
                 let data = vec![1u16; num_elements.try_into().unwrap()];
@@ -101,20 +91,13 @@ fn array_read_all_sharded(c: &mut Criterion) {
         group.bench_with_input(BenchmarkId::from_parameter(size), size, |b, &size| {
             // Write the data
             let store = zarrs::storage::store::MemoryStore::new();
-            let sharding_codec = Arc::new(
-                ShardingCodecBuilder::new(
-                    vec![32; 3].try_into().unwrap(),
-                    &zarrs::array::DataType::UInt8,
-                )
-                .build(),
-            );
             let array = zarrs::array::ArrayBuilder::new(
                 vec![size; 3],
                 vec![size; 3],
                 zarrs::array::DataType::UInt8,
                 1u8,
             )
-            .array_to_bytes_codec(sharding_codec)
+            .subchunk_shape(vec![32; 3])
             .build(store.into(), "/")
             .unwrap();
             let data = vec![0u8; num_elements.try_into().unwrap()];